Optimize reports
Avoiding unnecessary DB requests to find the list of documents where client and service is the same subsystem. This info is computable from document data.
Additionally, fixing invalid duplicate detection when client and producer requestInTs are in different report periods. "get_faulty_documents" did not find duplicates in that case.